rec: move Rec to rec.rs
