+++ rmm/monitor/src/rmi/rec/mod.rs
@@ -1,7 +1,9 @@
 mod params;
+pub mod rec;
 pub mod run;
 
 use self::params::Params;
+pub use self::rec::Rec;
 use self::run::Run;
 use super::realm::Rd;
 use crate::event::{realmexit, Context, Mainloop, RsiHandle};
@@ -14,60 +16,6 @@ use crate::rmm::granule::GranuleState;
 use core::mem::ManuallyDrop;
 
 extern crate alloc;
-
-pub struct Rec {
-    pub rd: &'static Rd,
-    vcpuid: usize,
-    ripas: Ripas,
-}
-
-struct Ripas {
-    start: u64,
-    end: u64,
-    addr: u64,
-    state: u8,
-}
-
-impl Rec {
-    pub unsafe fn new(
-        rec_addr: usize,
-        vcpuid: usize,
-        rd: &'static Rd,
-    ) -> ManuallyDrop<&'static mut Rec> {
-        let rec: &mut Rec = &mut *(rec_addr as *mut Rec);
-        rec.vcpuid = vcpuid;
-        rec.rd = rd;
-        ManuallyDrop::new(rec)
-    }
-
-    pub unsafe fn into(rec_addr: usize) -> ManuallyDrop<&'static mut Rec> {
-        let rec: &mut Rec = &mut *(rec_addr as *mut Rec);
-        ManuallyDrop::new(rec)
-    }
-
-    pub fn id(&self) -> usize {
-        self.vcpuid
-    }
-
-    pub fn set_ripas(&mut self, start: u64, end: u64, addr: u64, state: u8) {
-        self.ripas.start = start;
-        self.ripas.end = end;
-        self.ripas.addr = addr;
-        self.ripas.state = state;
-    }
-
-    pub fn inc_ripas_addr(&mut self, size: u64) {
-        self.ripas.addr += size;
-    }
-
-    pub fn ripas_addr(&mut self) -> usize {
-        self.ripas.addr as usize
-    }
-}
-
-impl Drop for Rec {
-    fn drop(&mut self) {}
-}
 
 pub fn set_event_handler(mainloop: &mut Mainloop) {
     listen!(mainloop, rmi::REC_CREATE, |arg, ret, rmm| {
